# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sia - Agent Orchestration API",
    "description": "Sia connects with code generation agents and receives your requests from Slack, Discord, or through Sia mobile app. It picks up the task, schedules it, and gets it done like your personal dev assistant.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": ["error"]
      },
...
      "DisconnectProviderResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": ["message"]
      },
...
      "StoreIntegrationSecretRequest": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "providerType": {
            "description": "The type of integration provider (e.g., cursor, claude-code, kiro-cli)",
            "type": "string"
          },
          "name": {
            "description": "A short description/label to help identify the key",
            "type": "string"
          },
          "apiKey": {
            "description": "The API key to store",
            "type": "string"
          }
        },
        "required": ["providerType", "name", "apiKey"]
      },
      "StoreIntegrationSecretResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "providerType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "storageType": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["gcp"]
              },
              {
                "type": "string",
                "enum": ["encrypted_local"]
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "providerType", "name", "storageType", "createdAt"]
      },
      "GetIntegrationSecretResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "providerType": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "storageType": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["gcp"]
              },
              {
                "type": "string",
                "enum": ["encrypted_local"]
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "providerType",
          "name",
          "storageType",
          "createdAt",
          "updatedAt"
        ]
      },
      "GetIntegrationSecretPlaintextResponse": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string"
          }
        },
        "required": ["apiKey"]
      },
...
    }
  },
  "paths": {
...
    "/integrations/secrets": {
      "post": {
        "tags": ["integrations"],
        "description": "Store an integration secret (API key)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoreIntegrationSecretRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Secret stored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreIntegrationSecretResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["integrations"],
        "description": "List all integration secrets for the current organization",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "providerType",
            "required": false,
            "description": "Filter by provider type (e.g., cursor, claude-code, kiro-cli)"
          }
        ],
        "responses": {
          "200": {
            "description": "Integration secrets retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetIntegrationSecretResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/secrets/{id}": {
      "get": {
        "tags": ["integrations"],
        "description": "Retrieve an integration secret",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The integration ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Integration secret retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegrationSecretResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["integrations"],
        "description": "Delete an integration secret (API key)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The integration ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Integration secret deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisconnectProviderResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/secrets/{id}/plaintext": {
      "get": {
        "tags": ["integrations"],
        "description": "Retrieve the plaintext API key for an integration (use with caution)",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The integration ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Plaintext secret retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIntegrationSecretPlaintextResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
...
}